use of size_t type in safety.h's TOKEN_LENGTH

+++ common/safety/safety.h
@@ -1,6 +1,8 @@
 #ifndef SAFETY_H
 #define SAFETY_H
 
-const int TOKEN_LENGTH = 6;
+#include <cstddef>
+
+const size_t TOKEN_LENGTH = 6;
 
 #endif // SAFETY_H
